src/main/java/xin/leiguang/sell/service/impl/CategoryServiceImpl.java
package xin.leiguang.sell.service.impl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import xin.leiguang.sell.dataobject.ProductCategory;
import xin.leiguang.sell.repository.ProductCategoryRepository;
import xin.leiguang.sell.service.CategoryService;

import java.util.List;

@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findById(categoryId).orElse(null);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> catetoryTypeList) {
        return repository.findByCategoryTypeIn(catetoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
